+++ src/main/java/persistence/sql/ddl/H2DropQueryBuilder.java
@@ -1,7 +1,7 @@
 package persistence.sql.ddl;
 
 import jakarta.persistence.Entity;
-import persistence.sql.ddl.model.TableName;
+import persistence.sql.model.TableName;
 
 public class H2DropQueryBuilder implements DropQueryBuilder {
